pyproject.toml
[project]
name = "abq-serpentine-substrate"
version = "0.1.0"
description = "ABAQUS 蛇形导线-基底界面力学仿真自动化脚本"
readme = "README.md"
requires-python = "==3.10.*"
# 后处理依赖
dependencies = [
    "hydra-core==1.3.2",
    "hydra-list-sweeper==1.1.0",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "pandas==2.3.3",
    "pydantic==2.12.5",
    "scipy==1.15.3",
]

[dependency-groups]
dev = [
    "abqpy>=2024.9,<2025.0", # 类型提示
    "pytest>=8.0.0",
]

[tool.uv]
add-bounds = "exact"

[build-system]
requires = ["setuptools>=77.0.3", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

# 命令行入口点 - 安装后可在任意位置运行
# 注意：ABAQUS 脚本（如 build_solid_serpentine.py）需要通过 abaqus cae noGUI= 运行
[project.scripts]
abq-gen-params = "abq_serp_sub.cli:generate_params"
